+++ Staging/Functions/Enable-OrchPersonalWorkspace.ps1
@@ -3,12 +3,12 @@ function Enable-OrchPersonalWorkspace {
     Param (
         [Parameter(Position = 0, Mandatory = $true, ValueFromPipelineByPropertyName = $true)]
         [SupportsWildcards()]
-        [ArgumentCompleter([UiPath.PowerShell.Completer.TenantUserUserNameCompleter[UiPath.PowerShell.Positional.UserName]])]
+        [ArgumentCompleter([UiPath.PowerShell.Completer.TenantUserUserNameCompleter])]
         [SupportsWildcards()]
         [string[]]$UserName,
 
         [Parameter(ValueFromPipelineByPropertyName = $true)]
-        [ArgumentCompleter([UiPath.PowerShell.Completer.DriveCompleter[UiPath.PowerShell.Positional.UserName]])]
+        [ArgumentCompleter([UiPath.PowerShell.Completer.DriveCompleter])]
         [string[]]$Path
     )